Remove AIDAPT imports and update make() call
The run_graph() function now takes a list of configuration keyword arguments that are passed to make. This should hopefully make the code more generic so that it can work regardless of the __init__() kwargs required to make a new registered object.
Also removed AIDAPT imports so that it makes sense for the core.