Create graph example workflow
This workflow does everything the current example_workflow does, except it does not output images. This is because the logdir is currently passed to the trainer and analysis modules as an input argument which is out of order. We may need to update the graph driver to allow for kwargs, but for now, this is a good example of what the workflow could look like.